Rework data fetching and calculations
Trends and counts are now computed on the workers and stored in the db.
UI no longer does any calculations - it just displays and sorts
precomputed data.

Data is now fetched serially, catching up on missing events with a 2 day
overlap into that past (in case there's something we've missed).

Then, fetched data is processed in parallel by the workers.